requirements.txt
flask==2.0.1
dash==2.0.0
pandas==1.3.3
requests==2.26.0
feedparser==6.0.8
pymongo==3.12.0
python-dotenv==0.19.0
spacy==3.1.3
anthropic==0.2.8
plotly==5.3.1
flask-login==0.5.0
schedule==1.1.0
flask-login==0.6.2